refactor(cli): share tsgolint path resolution
