+++ packages/cli/src/resolve-lint.ts
@@ -11,35 +11,12 @@
  * provides ESLint-compatible linting with significantly better performance.
  */
 
-import { existsSync, realpathSync } from 'node:fs';
 import { dirname, join } from 'node:path';
-import { fileURLToPath } from 'node:url';
 
 import { DEFAULT_ENVS, resolve } from './utils/constants.ts';
+import { resolveTsgolintExecutable } from './utils/tsgolint-path.ts';
 
-export function resolveWindowsTsgolintExecutable(
-  pathCandidates: string[],
-  options: {
-    exists: (path: string) => boolean;
-    getRealpathCandidates?: () => string[];
-  },
-): string {
-  let oxlintTsgolintPath = pathCandidates.find((p) => options.exists(p)) ?? '';
-  if (!oxlintTsgolintPath && options.getRealpathCandidates) {
-    try {
-      oxlintTsgolintPath = options.getRealpathCandidates().find((p) => options.exists(p)) ?? '';
-    } catch {
-      // realpath failed, fall through to default
-    }
-  }
-  if (!oxlintTsgolintPath) {
-    throw new Error(
-      'Unable to resolve oxlint-tsgolint executable, tried:\n' +
-        pathCandidates.map((path) => `- ${path}`).join('\n'),
-    );
-  }
-  return oxlintTsgolintPath;
-}
+export { resolveWindowsTsgolintExecutable } from './utils/tsgolint-path.ts';
 
 /**
  * Resolves the oxlint binary path and environment variables.
@@ -63,33 +40,10 @@ export async function lint(): Promise<{
   const oxlintMainPath = resolve('oxlint');
   const oxlintPackageRoot = dirname(dirname(oxlintMainPath));
   const binPath = join(oxlintPackageRoot, 'bin', 'oxlint');
-  let oxlintTsgolintPath = resolve('oxlint-tsgolint/bin/tsgolint');
-  if (process.platform === 'win32') {
-    // On Windows, try .exe first (bun creates .exe), then .cmd (npm/pnpm/yarn create .cmd)
-    const scriptDir = dirname(fileURLToPath(import.meta.url));
-    const localBinDir = join(scriptDir, '..', 'node_modules', '.bin');
-    const oxlintTsgolintPackagePath = dirname(dirname(oxlintTsgolintPath));
-    const projectBinDir = join(oxlintTsgolintPackagePath, '..', '.bin');
-    const pathCandidates = [
-      join(localBinDir, 'tsgolint.exe'),
-      join(localBinDir, 'tsgolint.cmd'),
-      join(projectBinDir, 'tsgolint.exe'),
-      join(projectBinDir, 'tsgolint.cmd'),
-    ];
-    oxlintTsgolintPath = resolveWindowsTsgolintExecutable(pathCandidates, {
-      exists: existsSync,
-      // Bun stores packages in .bun/ cache dirs where the symlinked paths above won't match.
-      getRealpathCandidates: () => {
-        const realPkgDir = realpathSync(join(scriptDir, '..'));
-        const realBinDir = join(dirname(realPkgDir), '.bin');
-        return [join(realBinDir, 'tsgolint.exe'), join(realBinDir, 'tsgolint.cmd')];
-      },
-    });
-    // Keep the resolved absolute path. oxlint may be spawned with a different cwd than
-    // this launcher (e.g. the workspace package dir under `vp run -r`), where a path made
-    // relative to the launcher's process.cwd() would resolve against the wrong base
-    // directory and fail (e.g. pnpm's `.pnpm` only exists at the monorepo root).
-  }
+  const oxlintTsgolintPath = resolveTsgolintExecutable(
+    resolve('oxlint-tsgolint/bin/tsgolint'),
+    import.meta.url,
+  );
   const result = {
     binPath,
     // TODO: provide envs inference API
